backend/backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
stripe
pytest
pytest-django
black
flake8
mypy
python-dotenv